--- src/shared/ui/bottom-sheet/index.ts
@@ -1,1 +1,0 @@
-export { BottomSheetModal } from './BottomSheetModal';
