+++ src/shared/ui/modal/index.ts
@@ -1,2 +1,3 @@
 export { ConfirmModal } from './ConfirmModal';
 export { ConfirmReportModal } from './ConfirmReportModal';
+export { BottomSheetModal } from './BottomSheetModal';
